Match integer types of manifest more closely
Generally, `partition_manifest` uses signed types for its metadata. The
decode does the opposite and uses unsigned types. For optional fields
this is problematic as they default to the min value of `int64_t`. When
decoding as unsigned the result is different which trips up tests.

+++ src/bucket_reader.rs
@@ -324,7 +324,7 @@ impl BucketReader {
             if let Some(manifest_segments) = &partition_manifest.segments {
                 for (segment_short_name, segment) in manifest_segments {
                     if let Some(so) = partition_manifest.start_offset {
-                        if segment.committed_offset < so {
+                        if segment.committed_offset < so as u64 {
                             debug!(
                                 "Not checking {} {}, it is below start offset",
                                 partition_manifest.ntp(),
